+++ CustomerChargeNotification/Models/CustomerGameCharge.cs
@@ -6,6 +6,6 @@ public class CustomerGameCharge
     public int CustomerId { get; set; }
     public required string GameName { get; set; }
     public required string Description { get; set; }
-    public double TotalCost { get; set; }
+    public int TotalCost { get; set; }
     public DateTime ChargeDate { get; set; }
 }
